pyproject.toml
[tool.poetry]
name = "electronic-retail-chain"
version = "0.1.0"
description = ""
authors = ["Julia Bengert <yubengert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework = "^3.16.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.1.0"
djangorestframework-simplejwt = "^5.5.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
isort = "^6.0.1"
mypy = "^1.15.0"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
line_length = 119


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
